+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/endpoints/NoMetricsSingleDimensionDataServletSpec.groovy
@@ -77,7 +77,12 @@ class NoMetricsSingleDimensionDataServletSpec extends BaseDataServletComponentSp
                 "type" : "table"
             },
             "dimensions": [
-                "model"
+                "model",
+                {
+                    "dimension":"misc",
+                    "outputName":"other",
+                    "type":"default"
+                }
             ],
             "filter": {
                 "fields": [